fix: unable to import kaggle data in jupyter notebooks
The Kaggle data is currently not being imported in Jupyter Notebooks due to a "403 Forbidden" error message.
We have been using the following Kaggle URL: https://www.kaggle.com/c/bluebook-for-bulldozers/overview.

The code snippet shows the Kaggle dataset definition and destination folder specification for downloading the dataset.

```python
KaggleDatasetPath = "c/bluebook-for-bulldozers/overview"
DestinationFolder = "inputs/datasets/raw"
! kaggle datasets download -d {KaggleDatasetPath} -p {DestinationFolder}

```

This issue needs to be resolved for successful data download and subsequent analysis.

On branch main
Your branch is up to date with 'origin/main'.

Changes to be committed:
	new file:   jupyter_notebooks/01_data_collection.ipynb